pyproject.toml
[tool.poetry]
name = "assistant"
version = "0.1.0"
description = ""
authors = ["Natalia <tsepeleva.natalia16@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest-playwright = "^0.5.2"
playwright = "^1.48.0"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
